+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Primary analysis experiment type in AppBio Quantstudio Design and Analysis
+- Added ThermoFisher Genesys 30 adapter
+- Added requirement doc for ThermoFisher Genesys 30 adapter
 - Added support for xlsx exports in Unchained Lunatic parser
 - Add column special character normalization to pandas util (and so most parsers)
 
@@ -18,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix SoftMax Pro case where some columns were incorrectly identified as not numeric for calculated data documents. Added exceptions for Masked and Range values which will be treated as NaN.
 
 ### Changed
+- Updated the spectrophotometry mapper script to accommodate absorbance spectrum data cubes
 
 ### Deprecated